Fixed the previous problem
In the previous commit, I mistakenly merged the shells of one solid. Now
this is reverted and a legacy bug in VSolid is found and partially
fixed. This bug will cause the output shells of a solid be identical,
because in an iteration, the Array is flushed in each loop.